phase4: Spike cluster verification with HTIF exit
Bare-metal 2-hart Spike pipeline for Phase 1 Rust SPSC IPC:

Build pipeline:
- Makefile: cross-compile rv32imac, link with MemPool crt0.S, run on Spike
- spike_link.ld: L1 TCDM at 0x10000, L2 at 0x80000000, .l1 NOLOAD sections
- spike_globals.S: ipc_queue (0x10080 bytes) + boot_barrier in .l1
- spike_htif.S: HTIF tohost/fromhost for Spike termination

HTIF exit protocol:
- Added tohost/fromhost externs to main.rs
- Core 0 writes (exit_code << 1) | 1 to tohost after final barrier
- Fixes infinite restart loop: _eoc -> boot ROM -> _start -> main()

Makefile fixes:
- timeout --foreground (dash process group breaks Spike HTIF polling)
- 120-second timeout for multi-hart init loop

Refactored src/lib.rs -> src/main.rs (bare-metal entry point split).
Updated primitives.rs hardware_spin_wait for Spike fallback path.
