+++ software/mempool-ipc/phase1_rust/.gitignore
@@ -1,2 +1,5 @@
 # Cargo build output
 target/
+
+# Generated linker script (preprocessed from arch.ld.c)
+spike_arch.ld
