+++ software/mempool-ipc/phase1_rust/Cargo.toml
@@ -1,4 +1,5 @@
 [package]
+autobins = false
 name = "mempool-ipc-phase1"
 version = "0.1.0"
 edition = "2021"
