+++ software/mempool-ipc/phase1_rust/build.rs
@@ -54,4 +54,5 @@ fn main() {
     ] {
         println!("cargo:rerun-if-env-changed={var}");
     }
+
 }
