+++ software/mempool-ipc/phase1_rust/spike_htif.S
@@ -1,0 +1,9 @@
+// HTIF interface for Spike emulator (tohost/fromhost communication)
+.section .htif, "aw", @progbits
+.globl tohost
+.globl fromhost
+.align 6
+tohost:   .word 0
+          .word 0
+fromhost: .word 0
+          .word 0
